Added UI handling of the three different cases of API response.

+++ app/src/main/java/com/cs4520/assignment1/ProductData.kt
@@ -10,4 +10,4 @@ data class ProductData(
     @SerializedName("expiryDate")
     val expiryDate: String?,
     @SerializedName("price")
-    val price: Int)
+    val price: Double)
